extract fs-observable out to a package
