+++ packages/fs-observable/package.json
@@ -1,0 +1,36 @@
+{
+  "name": "fs-observable",
+  "version": "1.0.0",
+  "description": "Observables for node fs",
+  "main": "lib/index.js",
+  "nteractDesktop": "src/index.js",
+  "scripts": {
+    "prepare": "npm run build",
+    "build": "npm run build:clean && npm run build:lib && npm run build:flow",
+    "build:clean": "rimraf lib",
+    "build:flow": "flow-copy-source -v -i '**/__tests__/**' src lib",
+    "build:lib": "babel -d lib src --ignore '**/__tests__/**'",
+    "build:lib:watch": "npm run build:lib -- --watch"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/nteract/nteract.git"
+  },
+  "keywords": ["rxjs", "fs", "observable"],
+  "author": "Kyle Kelley <rgbkrk@gmail.com>",
+  "license": "BSD-3-Clause",
+  "bugs": {
+    "url": "https://github.com/nteract/nteract/issues"
+  },
+  "homepage":
+    "https://github.com/nteract/nteract/tree/master/packages/fs-observable",
+  "publishConfig": {
+    "access": "public"
+  },
+  "devDependencies": {
+    "rxjs": "^5.4.1"
+  },
+  "peerDependencies": {
+    "rxjs": "^5.4.1"
+  }
+}
